% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conexion.R
\name{t.conectarbd}
\alias{t.conectarbd}
\title{Conecta R con la base de datos de analitica}
\usage{
\method{t}{conectarbd}(odbc = "analitica", user = "topaz", pass)
}
\arguments{
\item{odbc}{nombre de odbc local para conectar}
}
\description{
Usa paquete RODBC.
}

