% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacidades.R
\name{t_capacidades}
\alias{t_capacidades}
\title{Obtiene datos de capacidades por Sede UNID}
\usage{
t_capacidades(con, actual = TRUE, negocio = "UNID", version = 1)
}
\arguments{
\item{con}{nombre de la conexion (ver t_conectarbd())}

\item{actual}{si TRUE, obtiene la fecha más actual de actualización por campus}

\item{negocio}{default = UNID}
}
\description{
Base de Datos de Sedes UNID
}

