% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo.R
\name{t_censo2010}
\alias{t_censo2010}
\title{Obtiene datos del Censo de INEGI 2010}
\usage{
t_censo2010(con)
}
\arguments{
\item{con}{nombre de la conexion (ver t_conectarbd())}
}
\description{
Llama a base en SQL
}
