% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{t_datos_modelo}
\alias{t_datos_modelo}
\title{Descargar datos de Modelo de Expansion}
\usage{
t_datos_modelo(con, version, pruebaMty = FALSE)
}
\arguments{
\item{con}{conexion abierta de RODBC}

\item{version}{version de descarga}

\item{pruebaMty}{si TRUE, solamente descarga datos de Monterrey}
}
\description{
Descargas datos para modelar de sql.
}
