% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo.R
\name{t_descargar_censoecon}
\alias{t_descargar_censoecon}
\title{Obtiene datos del Censo Economico}
\usage{
t_descargar_censoecon(con)
}
\arguments{
\item{con}{nombre de la conexion (ver t_conectarbd())}
}
\description{
Descarga censo economico
}
