% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoecon.R
\name{t_descargar_censoecon}
\alias{t_descargar_censoecon}
\title{Descarga Censo Económico}
\usage{
t_descargar_censoecon(y = 2004, nivel = "MUNICIPIO", canal = tpz_con)
}
\arguments{
\item{y}{year}

\item{nivel}{nivel MUNICIPIO o ESTADO}

\item{canal}{canal de conexion a R}
}
\description{
Usa paquete RODBC, ver versiones.
}
